refactor(fetcher): Rename InterceptorManager to InterceptorRegistry
- Move FetcherInterceptors class to new file fetcherInterceptors.ts
- Rename InterceptorManager to InterceptorRegistry
- Update imports and references in affected files
