+++ packages/fetcher/test/interceptor.test.ts
@@ -13,14 +13,14 @@
 
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { Fetcher, FetcherInterceptors, FetchExchange, HttpMethod, Interceptor, InterceptorManager } from '../src';
+import { Fetcher, FetcherInterceptors, FetchExchange, HttpMethod, Interceptor, InterceptorRegistry, } from '../src';
 
 describe('interceptor.ts', () => {
   describe('InterceptorManager', () => {
-    let manager: InterceptorManager;
+    let manager: InterceptorRegistry;
 
     beforeEach(() => {
-      manager = new InterceptorManager();
+      manager = new InterceptorRegistry();
     });
 
     it('should have correct order value', () => {
@@ -242,9 +242,9 @@ describe('interceptor.ts', () => {
     it('should create interceptor managers for request, response and error', () => {
       const interceptors = new FetcherInterceptors();
 
-      expect(interceptors.request).toBeInstanceOf(InterceptorManager);
-      expect(interceptors.response).toBeInstanceOf(InterceptorManager);
-      expect(interceptors.error).toBeInstanceOf(InterceptorManager);
+      expect(interceptors.request).toBeInstanceOf(InterceptorRegistry);
+      expect(interceptors.response).toBeInstanceOf(InterceptorRegistry);
+      expect(interceptors.error).toBeInstanceOf(InterceptorRegistry);
     });
 
     it('should allow adding request interceptors', () => {
@@ -286,7 +286,7 @@ describe('interceptor.ts', () => {
 
   describe('Interceptor Types', () => {
     it('should process request interceptors correctly', async () => {
-      const manager = new InterceptorManager();
+      const manager = new InterceptorRegistry();
       const mockExchange = new FetchExchange(new Fetcher(), {
         url: 'http://example.com',
         method: HttpMethod.GET,
@@ -318,7 +318,7 @@ describe('interceptor.ts', () => {
     });
 
     it('should process response interceptors correctly', async () => {
-      const manager = new InterceptorManager();
+      const manager = new InterceptorRegistry();
       const response = new Response('{"data": "test"}', {
         status: 200,
         headers: { 'Content-Type': 'application/json' },
@@ -356,7 +356,7 @@ describe('interceptor.ts', () => {
     });
 
     it('should process error interceptors correctly', async () => {
-      const manager = new InterceptorManager();
+      const manager = new InterceptorRegistry();
       const error = new Error('Network error');
 
       const mockExchange = new FetchExchange(
